Pull out non-mutable interface for response
